TIC-77: Abstract away season api call to util folder
